requirements.txt
# 项目完整依赖配置 - 基于PaddleSeg和项目需求
# 基础依赖
numpy>=1.19.0
opencv-python>=4.5.0,<4.7.0
pillow>=8.0.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.3.0
seaborn>=0.11.0
pandas>=1.3.0
tqdm>=4.60.0
pyyaml>=5.4.0
requests>=2.25.0
filelock>=3.0.0
prettytable>=2.0.0

# PaddlePaddle相关
paddlepaddle-gpu==2.6.1
paddleseg>=2.8.0

# 图像处理
albumentations>=1.3.0
scikit-image>=0.19.0

# 可视化
visualdl>=2.5.0
plotly>=5.0.0

# GUI相关
PyQt5>=5.15.0
qtpy>=2.0.0

# 点云处理
open3d>=0.16.0

# 工具类
easydict>=1.9.0
wget>=3.2
cython>=0.29.0
protobuf==3.20.0

# 开发工具
jupyter>=1.0.0
ipykernel>=6.0.0